+++ src/app/manifestGeneration/manifestGenerators/generatorDevPreview/ManifestGenerator.ts
@@ -1,0 +1,27 @@
+// Copyright (c) Wictor Wilén. All rights reserved. 
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT license.
+
+import { BaseManifestGenerator } from "../../BaseManifestGenerator";
+import { TabManifestUpdater } from "./TabManifestUpdater";
+import { BotManifestUpdater } from "./BotManifestUpdater";
+import { ConnectorManifestUpdater } from "./ConnectorManifestUpdater";
+import { MessageExtensionManifestUpdater } from "./MessageExtensionManifestUpdater";
+import { GeneratorTeamsAppOptions } from "../../../GeneratorTeamsAppOptions";
+
+export class ManifestGenerator extends BaseManifestGenerator {
+    constructor() {
+        super();
+        this.tabUpdater = new TabManifestUpdater();
+        this.botUpdater = new BotManifestUpdater();
+        this.connectorUpdater = new ConnectorManifestUpdater();
+        this.messageExtensionUpdater = new MessageExtensionManifestUpdater();
+    }
+
+    public generateManifest(options: GeneratorTeamsAppOptions): any {
+        const manifest = super.generateManifest(options);
+        manifest["$schema"] = "https://raw.githubusercontent.com/OfficeDev/microsoft-teams-app-schema/preview/DevPreview/MicrosoftTeams.schema.json";
+        manifest.manifestVersion = "devPreview";
+        return manifest;
+    }
+}