+++ src/connector/templates/src/app/scripts/{connectorName}/__tests__/{connectorComponentName}Config.spec.tsx
@@ -2,20 +2,22 @@ import * as React from "react";
 import { shallow } from "enzyme";
 import toJson from "enzyme-to-json";
 
-import { <%=tabReactComponentName%>Config } from "../<%=tabReactComponentName%>Config";
+import { <%=connectorComponentName%>Config } from "../<%=connectorComponentName%>Config";
 
 describe("<%=tabReactComponentName%>Config Component", () => {
     // Snapshot Test Sample
     it("should match the snapshot", () => {
-        const wrapper = shallow(<<%=tabReactComponentName%>Config />);
+        window.alert = jest.fn();
+        const wrapper = shallow(<<%=connectorComponentName%>Config />);
         expect(toJson(wrapper)).toMatchSnapshot();
     });
 
     // Component Test Sample
     it("should render the tab", () => {
-        const component = shallow(<<%=tabReactComponentName%>Config />);
-        const divResult = component.containsMatchingElement(<div>Configure your tab</div>);
+        const component = shallow(<<%=connectorComponentName%>Config />);
+        const divResult = component.containsMatchingElement(<div>Configure your Connector</div>);
 
         expect(divResult).toBeTruthy();
     });
 });
+