+++ src/messageExtension/MessageExtensionGenerator.ts
@@ -11,6 +11,7 @@ import Project, { Scope, Decorator, Node, ClassDeclaration } from "ts-morph";
 import * as ts from 'typescript';
 import * as path from 'path';
 import * as Guid from 'guid';
+import { ManifestGeneratorFactory } from '../app/manifestGeneration/ManifestGeneratorFactory';
 
 
 let yosay = require('yosay');
@@ -221,38 +222,12 @@ export class MessageExtensionGenerator extends Generator {
 
     public writing() {
         if (this.options.messageExtension) {
+            const manifestGeneratorFactory = new ManifestGeneratorFactory();
+            const manifestGenerator = manifestGeneratorFactory.createManifestGenerator(this.options.manifestVersion);
             let manifestPath = "src/manifest/manifest.json";
             var manifest: any = this.fs.readJSON(manifestPath);
-
-            if (!manifest.composeExtensions) {
-                manifest.composeExtensions = [];
-            }
-
-            let command = {
-                id: this.options.messageExtensionName,
-                title: this.options.messageExtensionTitle,
-                description: 'Add a clever description here',
-                initialRun: true,
-                parameters: [
-                    {
-                        name: 'parameter',
-                        description: 'Description of the parameter',
-                        title: 'Parameter'
-                    }
-                ]
-            };
-
-            let composeExtension = manifest.composeExtensions.find((ce: { botId: string; }) => ce.botId == this.options.messageExtensionId);
-            if (this.options.existingManifest && composeExtension) {
-                composeExtension.commands.push(command);
-            } else {
-                // no existing manifest
-                manifest.composeExtensions.push({
-                    botId: this.options.messageExtensionId,
-                    canUpdateConfiguration: true,
-                    commands: [command]
-                });
-            }
+ 
+            manifestGenerator.updateMessageExtensionManifest(manifest, this.options);
 
             this.fs.writeJSON(manifestPath, manifest);
 
@@ -265,6 +240,13 @@ export class MessageExtensionGenerator extends Generator {
                     "src/app/scripts/{messageExtensionName}/{messageExtensionClassName}Config.tsx",
                     "src/app/web/{messageExtensionName}/config.html",
                 );
+
+                if(this.options.unitTestsEnabled) {
+                    templateFiles.push(
+                        "src/app/scripts/{messageExtensionName}/__tests__/{messageExtensionClassName}Config.spec.tsx"
+                    );
+                }
+
                 templateFiles.forEach(t => {
                     this.fs.copyTpl(
                         this.templatePath(t),
@@ -288,14 +270,7 @@ export class MessageExtensionGenerator extends Generator {
                     `Automatically added for the ${this.options.messageExtensionName} message extension`,
                     this.fs
                 );
-                // Yotilities.insertImportDeclaration(
-                //     `src/app/${ this.options.botName }.ts`,
-                //     this.options.messageExtensionName,
-                //     `./${ this.options.messageExtensionName }`,
-                //     `Automatically added for the ${ this.options.messageExtensionName } message extension`,
-                //     this.fs
-                // );
-
+ 
                 // Dynamically insert the reference and hook it up to the Bot
                 const project = new Project();
                 const file = project.createSourceFile(`src/app/${this.options.botName}/${this.options.botClassName}.ts`, this.fs.read(`src/app/${this.options.botName}/${this.options.botClassName}.ts`), {