+++ src/tab/templates/src/app/scripts/{tabName}/__tests__/{tabReactComponentName}Remove.spec.tsx
@@ -1,10 +1,16 @@
 import * as React from "react";
-import { shallow, mount } from "enzyme";
+import { shallow } from "enzyme";
+import toJson from "enzyme-to-json";
 
 import { <%=tabReactComponentName%>Remove } from "../<%=tabReactComponentName%>Remove";
-import { ThemeStyle } from "msteams-ui-styles-core";
 
 describe("<%=tabReactComponentName%>Remove Component", () => {
+    // Snapshot Test Sample
+    it("should match the snapshot", () => {
+        const wrapper = shallow(<<%=tabReactComponentName%>Remove />);
+        expect(toJson(wrapper)).toMatchSnapshot();
+    });
+
     // Component Test Sample
     it("should render the tab", () => {
         const component = shallow(<<%=tabReactComponentName%>Remove />);